+++ .buildkite/jobs/pipeline.android_rn_76.yml
@@ -1,4 +1,4 @@
-  - label: ":android::detox: (New Arch) - RN .76 + Android: Tests app"
+  - label: ":android::detox: RN .76 + Android: Tests app"
     command:
       - "nvm install"
       - "./scripts/ci.android.sh"
