+++ .buildkite/pipeline_common.sh
@@ -7,7 +7,7 @@ cat .buildkite/jobs/pipeline.ios_rn_76.yml
 cat .buildkite/jobs/pipeline.ios_rn_73.yml
 cat .buildkite/jobs/pipeline.ios_demo_app_rn_76_new_arch.yml
 cat .buildkite/jobs/pipeline.android_rn_76.yml
-cat .buildkite/jobs/pipeline.android_rn_75.yml
+cat .buildkite/jobs/pipeline.android_rn_76_old_arch.yml
 cat .buildkite/jobs/pipeline.android_rn_73.yml
 cat .buildkite/jobs/pipeline.android_demo_app_rn_76.yml
 cat .buildkite/pipeline.post_processing.yml
