+++ detox/android/detox/build.gradle
@@ -150,6 +150,9 @@ dependencies {
     api('androidx.test:core-ktx:1.6.1') {
         because 'Needed by Detox but also makes AndroidX test core seamlessly provided to Detox users with hybrid apps/E2E-tests.'
     }
+    api("org.jetbrains.kotlin:kotlin-reflect:$_kotlinVersion") {
+        because('Needed by Detox for kotlin reflection')
+    }
 }
 
 // Third-party/extension deps.
@@ -202,7 +205,6 @@ if (rootProject.hasProperty('isOfficialDetoxLib') ||
     dependencies {
         testImplementation 'org.spekframework.spek2:spek-dsl-jvm:2.0.15'
         testImplementation 'org.spekframework.spek2:spek-runner-junit5:2.0.15'
-        implementation "org.jetbrains.kotlin:kotlin-reflect:$_kotlinVersion"
     }
 }
