+++ detox/android/detox/src/full/java/com/wix/detox/reactnative/idlingresources/factory/FabricDetoxIdlingResourceFactoryStrategy.kt
@@ -1,7 +1,7 @@
 package com.wix.detox.reactnative.idlingresources.factory
 
 import com.facebook.react.ReactApplication
-import com.wix.detox.reactnative.getCurrentReactContextSafe
+import com.wix.detox.reactnative.getCurrentReactContext
 import com.wix.detox.reactnative.idlingresources.DetoxIdlingResource
 import com.wix.detox.reactnative.idlingresources.animations.AnimatedModuleIdlingResource
 import com.wix.detox.reactnative.idlingresources.network.NetworkIdlingResource
@@ -16,7 +16,7 @@ class FabricDetoxIdlingResourceFactoryStrategy(private val reactApplication: Rea
     override suspend fun create(): Map<IdlingResourcesName, DetoxIdlingResource> =
         withContext(Dispatchers.Main) {
             val reactContext =
-                reactApplication.getCurrentReactContextSafe()
+                reactApplication.getCurrentReactContext()
                     ?: throw IllegalStateException("ReactContext is null")
 
             val result = mutableMapOf<IdlingResourcesName, DetoxIdlingResource>(
