+++ detox/android/detox/src/full/java/com/wix/detox/reactnative/idlingresources/timers/FabricTimersIdlingResource.kt
@@ -3,11 +3,7 @@ package com.wix.detox.reactnative.idlingresources.timers
 import android.view.Choreographer
 import androidx.test.espresso.IdlingResource
 import com.facebook.react.bridge.ReactContext
-import com.facebook.react.modules.core.JavaTimerManager
 import com.wix.detox.reactnative.idlingresources.DetoxIdlingResource
-import org.joor.Reflect
-import kotlin.reflect.full.declaredFunctions
-import kotlin.reflect.jvm.isAccessible
 
 class FabricTimersIdlingResource(
     private val reactContext: ReactContext,
@@ -15,7 +11,7 @@ class FabricTimersIdlingResource(
 ) : DetoxIdlingResource(), Choreographer.FrameCallback {
 
     override fun checkIdle(): Boolean {
-        val hasActiveTimers: Boolean = hasActiveTimers()
+        val hasActiveTimers = JavaTimersReflected.hasActiveTimers(reactContext)
 
         if (hasActiveTimers) {
             getChoreographer().postFrameCallback(this@FabricTimersIdlingResource)
@@ -24,15 +20,6 @@ class FabricTimersIdlingResource(
         }
 
         return !hasActiveTimers
-    }
-
-    private fun hasActiveTimers(): Boolean {
-        val timersManager = getTimersManager()
-        val hasActiveTimersInRangeInstanceClass = timersManager::class
-        val method = hasActiveTimersInRangeInstanceClass.declaredFunctions.first { it.name.contains("hasActiveTimersInRange") }
-        method.isAccessible = true
-        val hasActiveTimers: Boolean = method.call(timersManager, BUSY_WINDOW_THRESHOLD) as? Boolean ?: false
-        return hasActiveTimers
     }
 
     override fun registerIdleTransitionCallback(callback: IdlingResource.ResourceCallback?) {
@@ -46,7 +33,7 @@ class FabricTimersIdlingResource(
     }
 
     override fun getDebugName(): String {
-        return "FabricTimers"
+        return "timers"
     }
 
     override fun getBusyHint(): Map<String, Any>? = null
@@ -55,11 +42,7 @@ class FabricTimersIdlingResource(
         return FabricTimersIdlingResource::class.java.name
     }
 
-    private fun getTimersManager(): JavaTimerManager {
-        val reactHost = Reflect.on(reactContext).field("mReactHost").get<Any>()
-        val reactInstance = Reflect.on(reactHost).field("mReactInstance").get<Any>()
-        return Reflect.on(reactInstance).field("mJavaTimerManager").get() as JavaTimerManager
-    }
+
 
     override fun doFrame(frameTimeNanos: Long) {
         isIdleNow()
